// src/app/users-list/create-edit-user/create-edit-user.component.ts
import { Component, Inject } from '@angular/core';
import {FormBuilder, FormGroup, ReactiveFormsModule, Validators} from '@angular/forms';
import {
  MAT_DIALOG_DATA,
  MatDialogActions,
  MatDialogClose,
  MatDialogContent,
  MatDialogRef, MatDialogTitle
} from '@angular/material/dialog';
import {MatError, MatFormField, MatLabel} from "@angular/material/form-field";
import {MatButton} from "@angular/material/button";
import {NgIf} from "@angular/common";
import {MatInput} from "@angular/material/input";
import {UsersService} from "../../services/usersService/users.service";
import {IUser} from "../../interfaces/iuser";
import { v4 as uuidv4 } from 'uuid';
@Component({
  selector: 'create-edit-user',
  templateUrl: './create-edit-user.component.html',
  standalone: true,
  imports: [
    MatFormField,
    MatDialogActions,
    MatDialogClose,
    MatButton,
    NgIf,
    MatInput,
    ReactiveFormsModule,
    MatDialogContent,
    MatDialogTitle,
    MatLabel,
    MatError
  ],
  styleUrls: ['./create-edit-user.component.scss']
})
export class CreateEditUserComponent {
  createEditForm: FormGroup;

  constructor(
    private fb: FormBuilder,
    private dialogRef: MatDialogRef<CreateEditUserComponent>,
    private usersService: UsersService,
    @Inject(MAT_DIALOG_DATA) public data: { user: IUser; isEdit: boolean }
  ) {
    this.createEditForm = this.fb.group({
      name: [data?.user?.name || '', Validators.required],
      email: [data?.user?.email || '', [Validators.required, Validators.email]],
      company: this.fb.group({
        name: [data?.user?.company?.name || '', Validators.required],
      })
    });
  }

  submitForm(): void {
    if (this.createEditForm.valid) {
      const user = this.createEditForm.value;

      if (this.data.isEdit && this.data.user.id) {
        user.id = this.data.user.id;
        this.usersService.updateUser(user);
      } else {
        user.id = uuidv4();
        this.usersService.addUser(user);
      }

      this.dialogRef.close();
    }
  }


}
